make watch for handlebars file
